requirements.txt
accelerate==0.25.0
deepspeed==0.12.5
einops==0.4.1
# flash-attn==2.3.6
huggingface-hub==0.20.1
imageio==2.33.1
lazy-loader==0.3
matplotlib==3.8.2
ninja==1.11.1.1
numpy==1.24.2
opencv-python==4.8.0.74
pandas==2.1.4
peft==0.5.0
pillow==9.4.0
pycocoevalcap==1.2
pycocotools==2.0.6
pyyaml==6.0.1
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.2
sentencepiece==0.1.99
shapely==2.0.4
tokenizers==0.13.3
tqdm==4.64.1
transformers==4.28.0
gradio==3.35.2
terminaltables==3.1.10